Try calculating score when checking if number in card, pop when hit
If a score is returned, that's a bingo!
Fancy try/except is cool with the custom Bingo exception, but meh..

No need to override the contains. Just have the method that check is
bingo handle if number is in the card, has a hit, and calculate score if
there's a bingo.

Pop from numbers dict, no need to keep track of what's been hit
separately. Easier to calculate the sum of unhits for the score!